Add files via upload
In this project employers are often interested in seeing a resume of candidates that they plan to hire. A resume can be presented in website form, as part of an online portfolio. In this project I build a resume web page that is responsive to different types of devices: mobile phones, tablets, and desktop machines.